move components/AmplitudeAnalytics to analytics/amplitude
