+++ src/components/swap/ConfirmSwapModal.tsx
@@ -1,11 +1,11 @@
 import { Trans } from '@lingui/macro'
 import { Trade } from '@uniswap/router-sdk'
 import { Currency, CurrencyAmount, Percent, Token, TradeType } from '@uniswap/sdk-core'
-import { sendAnalyticsEvent } from 'components/AmplitudeAnalytics'
-import { ModalName } from 'components/AmplitudeAnalytics/constants'
-import { EventName } from 'components/AmplitudeAnalytics/constants'
-import { Trace } from 'components/AmplitudeAnalytics/Trace'
-import { formatPercentInBasisPointsNumber, formatToDecimal, getTokenAddress } from 'components/AmplitudeAnalytics/utils'
+import { sendAnalyticsEvent } from 'analytics/amplitude'
+import { ModalName } from 'analytics/amplitude/constants'
+import { EventName } from 'analytics/amplitude/constants'
+import { Trace } from 'analytics/amplitude/Trace'
+import { formatPercentInBasisPointsNumber, formatToDecimal, getTokenAddress } from 'analytics/amplitude/utils'
 import { ReactNode, useCallback, useEffect, useMemo, useState } from 'react'
 import { InterfaceTrade } from 'state/routing/types'
 import { computeRealizedPriceImpact } from 'utils/prices'
