+++ src/components/swap/SwapModalHeader.tsx
@@ -1,9 +1,9 @@
 import { Trans } from '@lingui/macro'
 import { Currency, Percent, TradeType } from '@uniswap/sdk-core'
 import { Price } from '@uniswap/sdk-core'
-import { sendAnalyticsEvent } from 'components/AmplitudeAnalytics'
-import { EventName, SWAP_PRICE_UPDATE_USER_RESPONSE } from 'components/AmplitudeAnalytics/constants'
-import { formatPercentInBasisPointsNumber } from 'components/AmplitudeAnalytics/utils'
+import { sendAnalyticsEvent } from 'analytics/amplitude'
+import { EventName, SWAP_PRICE_UPDATE_USER_RESPONSE } from 'analytics/amplitude/constants'
+import { formatPercentInBasisPointsNumber } from 'analytics/amplitude/utils'
 import { RedesignVariant, useRedesignFlag } from 'featureFlags/flags/redesign'
 import { useEffect, useState } from 'react'
 import { AlertTriangle, ArrowDown } from 'react-feather'
