+++ src/lib/hooks/useApproval.ts
@@ -2,9 +2,9 @@ import { MaxUint256 } from '@ethersproject/constants'
 import type { TransactionResponse } from '@ethersproject/providers'
 import { Currency, CurrencyAmount, Token } from '@uniswap/sdk-core'
 import { useWeb3React } from '@web3-react/core'
-import { sendAnalyticsEvent } from 'components/AmplitudeAnalytics'
-import { EventName } from 'components/AmplitudeAnalytics/constants'
-import { getTokenAddress } from 'components/AmplitudeAnalytics/utils'
+import { sendAnalyticsEvent } from 'analytics/amplitude'
+import { EventName } from 'analytics/amplitude/constants'
+import { getTokenAddress } from 'analytics/amplitude/utils'
 import { useTokenContract } from 'hooks/useContract'
 import { useTokenAllowance } from 'hooks/useTokenAllowance'
 import { useCallback, useMemo } from 'react'
