+++ src/pages/App.tsx
@@ -1,6 +1,6 @@
-import { initializeAnalytics, sendAnalyticsEvent, user } from 'components/AmplitudeAnalytics'
-import { CUSTOM_USER_PROPERTIES, EventName, PageName } from 'components/AmplitudeAnalytics/constants'
-import { Trace } from 'components/AmplitudeAnalytics/Trace'
+import { initializeAnalytics, sendAnalyticsEvent, user } from 'analytics/amplitude'
+import { CUSTOM_USER_PROPERTIES, EventName, PageName } from 'analytics/amplitude/constants'
+import { Trace } from 'analytics/amplitude/Trace'
 import Loader from 'components/Loader'
 import TopLevelModals from 'components/TopLevelModals'
 import { useFeatureFlagsIsLoaded } from 'featureFlags'
