+++ src/pages/Pool/index.tsx
@@ -1,9 +1,9 @@
 import { Trans } from '@lingui/macro'
 import { useWeb3React } from '@web3-react/core'
-import { PageName } from 'components/AmplitudeAnalytics/constants'
-import { ElementName, Event, EventName } from 'components/AmplitudeAnalytics/constants'
-import { Trace } from 'components/AmplitudeAnalytics/Trace'
-import { TraceEvent } from 'components/AmplitudeAnalytics/TraceEvent'
+import { PageName } from 'analytics/amplitude/constants'
+import { ElementName, Event, EventName } from 'analytics/amplitude/constants'
+import { Trace } from 'analytics/amplitude/Trace'
+import { TraceEvent } from 'analytics/amplitude/TraceEvent'
 import { ButtonGray, ButtonPrimary, ButtonText } from 'components/Button'
 import { AutoColumn } from 'components/Column'
 import { FlyoutAlignment, NewMenu } from 'components/Menu'
