+++ src/pages/Vote/VotePage.tsx
@@ -2,8 +2,8 @@ import { BigNumber } from '@ethersproject/bignumber'
 import { Trans } from '@lingui/macro'
 import { CurrencyAmount, Fraction, Token } from '@uniswap/sdk-core'
 import { useWeb3React } from '@web3-react/core'
-import { PageName } from 'components/AmplitudeAnalytics/constants'
-import { Trace } from 'components/AmplitudeAnalytics/Trace'
+import { PageName } from 'analytics/amplitude/constants'
+import { Trace } from 'analytics/amplitude/Trace'
 import ExecuteModal from 'components/vote/ExecuteModal'
 import QueueModal from 'components/vote/QueueModal'
 import { NavBarVariant, useNavBarFlag } from 'featureFlags/flags/navBar'
