+++ src/state/transactions/updater.tsx
@@ -1,8 +1,8 @@
 import { Currency, Percent, TradeType } from '@uniswap/sdk-core'
 import { useWeb3React } from '@web3-react/core'
-import { sendAnalyticsEvent } from 'components/AmplitudeAnalytics'
-import { EventName } from 'components/AmplitudeAnalytics/constants'
-import { formatPercentInBasisPointsNumber, formatToDecimal, getTokenAddress } from 'components/AmplitudeAnalytics/utils'
+import { sendAnalyticsEvent } from 'analytics/amplitude'
+import { EventName } from 'analytics/amplitude/constants'
+import { formatPercentInBasisPointsNumber, formatToDecimal, getTokenAddress } from 'analytics/amplitude/utils'
 import { DEFAULT_TXN_DISMISS_MS, L2_TXN_DISMISS_MS } from 'constants/misc'
 import LibUpdater from 'lib/hooks/transactions/updater'
 import { useCallback, useMemo } from 'react'
